Changes stripe_helpers to account_helpers
