+++ app.rb
@@ -18,14 +18,16 @@ require 'active_support/number_helper'
   end
 end
 require 'jobs'
+require 'query_handler'
+require 'queries'
 
 require 'app/errors'
 require 'app/extensions'
 require 'app/helpers'
-require 'app/stripe_helpers'
+require 'app/account_helpers'
 require 'app/routes'
 
-include HuBoard::StripeHelpers
+include HuBoard::AccountHelpers
 
 module HuBoard
   class EnsureEMRunning
