+++ spec/spec_helper.rb
@@ -1,3 +1,1 @@
-Dir.glob('lib/**/*_spec*').each do |file|
-  require File.expand_path(file)
-end
+require "query_handler"
